updated code

+++ contracts/interface/IERC721Mint.sol
@@ -6,6 +6,6 @@ interface IERC721Mint is IERC721 {
     function mint(
         address to,
         string memory tokenURI,
-        uint256 _royality
+        uint96 _royality
     ) external returns (uint256);
 }
